+++ packages/project-access/src/project/index.ts
@@ -35,7 +35,8 @@ export {
     findProjectRoot,
     getAppRootFromWebappPath,
     findCapProjectRoot,
-    findRootsForPath
+    findRootsForPath,
+    isPathForCapApp
 } from './search';
 export { getWebappPath, readUi5Yaml } from './ui5-config';
 export { getMtaPath } from './mta';