+++ packages/project-access/test/project/cap.test.ts
@@ -21,6 +21,7 @@ import {
     isCapProject,
     deleteCapApp
 } from '../../src';
+import { isPathForCapApp } from '../../src/project/search';
 import * as file from '../../src/file';
 import os from 'os';
 import type { Logger } from '@sap-ux/logger';
@@ -86,6 +87,28 @@ describe('Test isCapJavaProject()', () => {
         expect(
             await isCapJavaProject(join(__dirname, '..', 'test-data', 'project', 'find-all-apps', 'CAP', 'CAPnode_mix'))
         ).toBeFalsy();
+    });
+});
+describe('Test isPathForCapApp()', () => {
+    test('Should return true for root CAP path', async () => {
+        const result = await isPathForCapApp(
+            join(__dirname, '../test-data/project/find-all-apps/CAP/CAPnode_fiori_elements/app/fiori_elements/')
+        );
+        expect(result).toBe(true);
+    });
+
+    test('Should return true for sub directory CAP path', async () => {
+        const result = await isPathForCapApp(
+            join(__dirname, '../test-data/project/find-all-apps/CAP/CAPnode_fiori_elements/app/fiori_elements/webapp/')
+        );
+        expect(result).toBe(true);
+    });
+
+    test('Should return false for non CAP path', async () => {
+        const result = await isPathForCapApp(
+            join(__dirname, '../test-data/project/find-all-apps/find-all-apps/adaptations/valid-adaptation')
+        );
+        expect(result).toBe(false);
     });
 });