+++ packages/preview-middleware/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/preview-middleware
 
+## 0.16.79
+
+### Patch Changes
+
+-   Updated dependencies [a64a3a5]
+    -   @sap-ux/project-access@1.27.5
+    -   @sap-ux/adp-tooling@0.12.54
+
 ## 0.16.78
 
 ### Patch Changes